refactor: 型の見直し
