+++ src/app/_app.unused.tsx
@@ -14,7 +14,11 @@ import { useIsInitializationMutators } from '@/store/isInitializationState'
 import { useNaturesMutators } from '@/store/naturesState'
 import { usePokemonBasicInfosSMutators } from '@/store/pokemonBasicInfosState'
 import { theme, useMediaQueryUp } from '@/utils/theme'
-import type { AuthUser, Nature, PokemonBasicInfo } from '@/types/front'
+import type {
+  AuthUser,
+  Nature,
+  PokemonBasicInfo,
+} from '@/types/openapi/schemas'
 import type { AxiosError } from 'axios'
 import type { AppProps } from 'next/app'
