+++ src/hooks/usePokemonStats.ts
@@ -11,7 +11,8 @@ import {
   numberToInt,
   valueVerification,
 } from '@/utils/helpers'
-import type { Nature, Pokemon, StatsKey } from '@/types/front'
+import type { Pokemon, StatsKey } from '@/types/front'
+import type { Nature } from '@/types/openapi/schemas'
 
 // ポケモンのステータス計算に関するカスタムフック
 export const usePokemonStats = (pokemon: Pokemon) => {
