+++ src/types/front/index.d.ts
@@ -1,50 +1,17 @@
-type NatureName =
-  | 'いじっぱり'
-  | 'うっかりや'
-  | 'おくびょう'
-  | 'おだやか'
-  | 'おっとり'
-  | 'おとなしい'
-  | 'がんばりや'
-  | 'きまぐれ'
-  | 'さみしがり'
-  | 'しんちょう'
-  | 'すなお'
-  | 'ずぶとい'
-  | 'せっかち'
-  | 'なまいき'
-  | 'てれや'
-  | 'のうてんき'
-  | 'のんき'
-  | 'ひかえめ'
-  | 'まじめ'
-  | 'むじゃき'
-  | 'やんちゃ'
-  | 'ゆうかん'
-  | 'ようき'
-  | 'れいせい'
-  | 'わんぱく'
+import type {
+  Nature,
+  PokemonBasicInfo,
+  PokemonSummary,
+  User,
+} from '@/types/openapi/schemas'
 
-export type Nature = Readonly<{
-  name: NatureName
-  increasedStat: StatsKey | null
-  decreasedStat: StatsKey | null
-}>
+type NatureName = Nature['name']
 
 type Rank = 'legendary' | 'mythical' | 'mega' | 'sv' | 'NotInPokedex'
 
 export type RankCheckbox = Readonly<{
   text: string
   value: Exclude<Rank, 'NotInPokedex'>
-}>
-
-export type PokemonBasicInfo = Readonly<{
-  no: number
-  name: string
-  form: string
-  ranks: ReadonlyArray<Rank>
-  evolutions: ReadonlyArray<number>
-  baseStats: Stats
 }>
 
 export type Pokemon = Readonly<{
@@ -55,38 +22,6 @@ export type Pokemon = Readonly<{
   evs: NullableStats
   description: string
 }>
-
-export type PokemonParams = {
-  name: string
-  nature: string
-  level: number | null
-  ivs: {
-    hp: number | null
-    attack: number | null
-    defense: number | null
-    spAttack: number | null
-    spDefense: number | null
-    speed: number | null
-  }
-  evs: {
-    hp: number | null
-    attack: number | null
-    defense: number | null
-    spAttack: number | null
-    spDefense: number | null
-    speed: number | null
-  }
-  realNumbers: {
-    hp: number
-    attack: number
-    defense: number
-    spAttack: number
-    spDefense: number
-    speed: number
-  }
-  description: string
-  isPublic: boolean
-}
 
 export type Stats = Readonly<{
   hp: number
@@ -104,27 +39,6 @@ type Nullable<T> = {
 }
 
 export type NullableStats = Nullable<Stats>
-
-export type LoginParams = {
-  email: string
-  password: string
-}
-
-export type User = Readonly<{
-  nickname: string
-  username: string
-}>
-
-export type AuthUser = User &
-  Readonly<{
-    id: number | null
-    email: string
-    isAuthenticated: boolean
-  }>
-
-export type Email = {
-  email: string
-}
 
 export type ShowUser = {
   user: User
